не работает. ошибка при авторизации. не могу понять в чем дело. также при запуске не показывается страница индекс

+++ src/main/java/JM/Task242/dao/UserDAO.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface UserDAO {
     List<User> allUsers();
     void add(User user);
-    void delete(int id);
+    void delete(Long id);
     void edit(User user);
-    User getById(int id);
+    User getById(Long id);
     User getUserByName(String name);
 }
